The first commit's message is:
Tuple value for radius

In confocal stacks, mostly the z pixel size is larger than the xy pixel
size. Features will have a different diameter and separation. This
feature allows for a tuple value of diameter and separation.

The binary_mask function could perform better, but it works.

The bpass filter does not allow different smoothing sizes, it just takes
the first value (depending on the order of your image, typically this is
the z value). This should be changed.

Also a change in subpx bias plot has been made so that it also plots the
z subpx bias.

This is the 2nd commit message:

Update masks.py

Unsure wether x,y order is right for 2D images (unimportant if feature size x = feature size y)

This is the 3nd commit message:

Corrected 1D code

This is the 4th commit message:

Updated bandpass filter for 3D

Function bandpass accepts tuple value for smoothing and noise. Noise
filtering is not done for dimensions with noise*2 <= smoothing.

Updated initiation at start of feature, added dimension mismatch check.