FIX: Redone feature Rg tests
Due to a factor 1/2 inside the original gaussian feature, Rg turned out
consequently larger than necessary. To compensate, the mask size was
chosen too small so that the tests pass.  The mask size should be very
large, so that the full region is integrated giving the right Rg.

As the factor 1/2 is removed, the mask size is now increased a lot to
get the appropriate Rg. This turns out to be consequently 0.2 larger,
which is curiously close to the 1/6 'mistake' in the Crocker IDL code.
However, adding the 1/6 would make the error only worse. It's probably
some pixelation thing why Rg is always overestimated. I chose to expect
0.2 higher, so that tests pass.