ENH Anisotropic feature characterization
A number of changes that reach deep into the feature finding: please
review carefully! The fixes are linked in subtle ways, that's why this
is one big commit.

- Changes the behaviour of the `mass` column: it is now calculated from
the raw image. This solves the issue #237 due to the rescaling of the
processed image and the subsequent calculation of mass on this image.
- Changes the behaviour of the `signal` column: previously it was
calculated from the raw image and subsequently the `black_level` was
subtracted. Because of the high sensitivity to noise, this gives
unpredictable results. It has been changed to: `signal =
preprocessed_feature.max()`. The rescaling has been accounted for via an
extra variable coming from `scale_to_gamut`.
- eccentricity is calculated on anisotropic masks too, assuming that
pixels are square
- size is calculated in all directions separately when an anisotropic
mask is used, giving `size_x`, `size_y` columns
- filtering on `maxsize` is done on sqrt(size_x**2 + size_y**2), which
equals the current behaviour if sizes are equal
- static error is calculated in all directions separately when an
anisotropic mask is used *or* when noise_size is anisotropic, giving
`ep_x`, `ep_y` columns
- `preprocessing.bandpass` now calculates the background level and
standard deviation (only if needed, defaults to current behaviour)
- `scale_to_gamut` returns the scale factor (only if needed, defaults to
current behaviour)
- docstring for eccentricity: 0 is circular, not 1 as it is documented
now
- characterization arrays are only initialized if needed