REF/API Change behaviour of draw_feature size arg
Formerly, the third argument of draw_feature, draw_spots and draw_array
determined the size of the box in which the feature was drawn. Now it
determines the size (radius of gyration) of the feature itself. The box
is determined by a different kwarg 'mask_diameter' and is by default
four times the size of the feature.