Fix rolling sum parameter in compute_drift.
This fixes an issue with how the rolling sum parameter was being used.

The issue:
The pandas rolling() function returns a Rolling object, which does not
have a cumulative sum function cumsum(). The cumsum() function is called
on the trajectory dataframe, which works fine if no rolling sum
parameter is passed, but if there is a rolling sum parameter passed,
the cumsum() function is being called on a Rolling object, which is
undefined, causing compute_drift() to crash with an exception
(Rolling object does not have a cumsum() function).

The fix:
This fixes the problem by calling sum() on the Rolling object,
which returns a dataframe. The cumsum() function can now be called
on the function.

To do:
Add a test for compute_drift() that uses a rolling sum parameter.
This problem would have been caught with more comprehensive test
coverage.