Remove dependency of test suite on nose.
Nose is quite unlikely to be installed on a dev's machine these days.

The test suite can fairly easily be converted to plain unittest, except
for the "slow" attrs which would need pytest to be replicated.  I'm not
against making the test suite depend on pytest (which is quite common
nowadays), but I'm not really sure it's worth it either -- the slow
tests don't add *so much* runtime to the test suite either.