FIX: scale up N to match the total N we had before
The effective number of measurements is difficult to calculate for cases of trajectories with gaps. Trackpy has previously used a scale factor len(traj) / len(pos) to adjust the effective number of measurements N. I have recently implemented an adjustment to this so that column values with no information will have an N of 0 (which is correct) instead of an estimated N. But setting some N values to zero makes the total N lower than it was before. I scale up the rest of the column so that the total N remains the same as it was before.

This is only approximate.