Match subnets in O(n^3) via linear_sum_assignment.
To solve subnets, instead of brute-forcing the n! combinations, use
scipy.optimize.linear_sum_assignment (commonly known as the
Hungarian/Munkres algorithm, although scipy actually uses a different
algorithm) which provides a solution in O(n^3).  This is not an original
idea; see e.g.
https://imagej.net/imagej-wiki-static/TrackMate_Algorithms.html#Solving_LAP

Locally, this method solves the "slow" example in adaptive-search.ipynb
(`tracks_regular = trackpy.link_df(cg, 0.75)`) in ~50ms instead of 1min.
linear_sum_assignment ("lsa") was extensively benchmarked in scipy
PR#12541 (which was most about adding a *sparse* variant, but one can
just look at the performance of lsa), which shows sub-second runtimes
for thousands of inputs.  This is also why this PR fully skips the use
of MAX_SUB_NET (at most, one may consider setting an alternate value in
the thousands for this strategy...).

The docs will need to be updated, but this PR is just to discuss the
implementation.